feat: hydrate Card component and update styling
