+++ src/pages/index.mdx
@@ -34,7 +34,6 @@ export const meta = {
   image={heroImage}
   imageAlt="CroCoder crocodile in a hoodie holding a laptop"
 />
-<Island deferUntil="visible">
 <OurClients
   title="Our Clients"
   text="At <em><b>Cro</b>Coder</em>, we’ve worked with various organizations, including enterprises, startups, and small businesses. We’ve built MVPs, helped projects scale up, refactored legacy code, and set up the architecture for transitioning into the world of React/Node. Is there something we can do for you?"
@@ -84,7 +83,7 @@ export const meta = {
       url: "https://www.misterspex.de/",
     },
   ]}
-/></Island>
+/>
 <HowWeWork
   title="How We Work"
   content={[
